# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: MemoryOps API
  version: 0.21.0
  description: Control plane API for AI-agent memory ingestion, lifecycle management, retrieval, and audit.
  contact:
    name: MemoryOps
    url: https://github.com/Quazmoz/memoryops
servers:
  - url: http://localhost:8080
security:
  - ApiKeyAuth: []
...
paths:
  /v1/ingest/github/{workspace_id}:
    post:
      tags: [Webhooks]
      operationId: ingestGitHub
      summary: Receive a GitHub webhook event
      description: Validates the workspace integration GitHub signature and accepts supported repository events for ingestion.
      security: []
      parameters:
        - $ref: '#/components/parameters/WebhookWorkspaceId'
        - name: X-Hub-Signature-256
          in: header
          required: true
          schema:
            type: string
        - name: X-GitHub-Event
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '202':
          description: Event accepted for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookAccepted'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
...
    WebhookWorkspaceId:
      name: workspace_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
...
    WebhookAccepted:
      type: object
      required: [event_id]
      properties:
        event_id:
          type: string
          format: uuid
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string